refactor: use int64/bigint for all IDs

+++ backend/auth/internal/handlers/auth.handler.go
@@ -53,7 +53,7 @@ func generateUniqueUsername(tx *gorm.DB, emailPrefix string) (string, error) {
 	return "", status.Error(codes.Internal, "failed to generate unique username after maximum retries")
 }
 
-func createSession(userID int) (*models.Session, error) {
+func createSession(userID int64) (*models.Session, error) {
 	sessionKey := uuid.New()
 	sessionExpiresAt := time.Now().Add(time.Duration(env.SESSION_EXPIRES_IN_HOURS) * time.Hour)
 
@@ -114,7 +114,7 @@ func (s *AuthServer) LoginWithPassword(ctx context.Context, req *proto.LoginWith
 	grpc.SetHeader(ctx, md)
 
 	return &proto.UserResponse{
-		Id:        int32(user.ID),
+		Id:        user.ID,
 		Username:  user.Username,
 		Email:     user.Email,
 		FirstName: user.FirstName.String,
@@ -277,7 +277,7 @@ func (s *AuthServer) VerifySignup(ctx context.Context, req *proto.VerificationRe
 	}
 
 	return &proto.UserResponse{
-		Id:        int32(user.ID),
+		Id:        user.ID,
 		Username:  user.Username,
 		Email:     user.Email,
 		FirstName: user.FirstName.String,
@@ -333,7 +333,7 @@ func (s *AuthServer) VerifyLoginOtp(ctx context.Context, req *proto.Verification
 	}
 
 	return &proto.UserResponse{
-		Id:        int32(user.ID),
+		Id:        user.ID,
 		Username:  user.Username,
 		Email:     user.Email,
 		FirstName: user.FirstName.String,
@@ -453,6 +453,6 @@ func (s *AuthServer) Authenticate(ctx context.Context, req *proto.AuthenticateRe
 		return nil, status.Error(codes.Unauthenticated, "invalid token claims")
 	}
 
-	userID := int32(claims["user_id"].(float64))
+	userID := int64(claims["user_id"].(float64))
 	return &proto.AuthenticateResponse{UserId: userID}, nil
 }
