+++ backend/auth/internal/handlers/user.handler.go
@@ -24,7 +24,7 @@ func (s *AuthServer) GetUser(ctx context.Context, req *proto.GetUserRequest) (*p
 	}
 
 	return &proto.UserProfileResponse{
-		Id:        int32(user.ID),
+		Id:        user.ID,
 		Username:  user.Username,
 		Email:     user.Email,
 		FirstName: user.FirstName.String,
@@ -72,7 +72,7 @@ func (s *AuthServer) UpdateUser(ctx context.Context, req *proto.UpdateUserReques
 
 	return &proto.UpdateUserResponse{
 		User: &proto.UserProfileResponse{
-			Id:        int32(user.ID),
+			Id:        user.ID,
 			Username:  user.Username,
 			Email:     user.Email,
 			FirstName: user.FirstName.String,
@@ -130,7 +130,7 @@ func (s *AuthServer) UpsertUser(ctx context.Context, req *proto.UpsertUserReques
 	}
 
 	return &proto.UserProfileResponse{
-		Id:        int32(user.ID),
+		Id:        user.ID,
 		Username:  user.Username,
 		Email:     user.Email,
 		FirstName: user.FirstName.String,
