+++ backend/common/pkg/models/exam_questions.go
@@ -1,10 +1,10 @@
 package models
 
 type ExamQuestion struct {
-	ID         int  `gorm:"primaryKey"`
-	ExamID     int  `gorm:"not null"`
-	QuestionID int  `gorm:"not null"`
-	Exam       Exam `gorm:"foreignKey:ExamID"`
+	ID         int64 `gorm:"primaryKey"`
+	ExamID     int64 `gorm:"not null"`
+	QuestionID int64 `gorm:"not null"`
+	Exam       Exam  `gorm:"foreignKey:ExamID"`
 }
 
 func (ExamQuestion) TableName() string {
