+++ backend/common/pkg/models/paper.go
@@ -12,7 +12,7 @@ type QuestionCount struct {
 }
 
 type Paper struct {
-	ID              int                `gorm:"primaryKey"`
+	ID              int64              `gorm:"primaryKey"`
 	Title           string             `gorm:"type:varchar(255);not null;default:'Untitled Paper'"`
 	MaxScore        int                `gorm:"default:0"`
 	DurationMinutes int                `gorm:"not null"`
